[core-amqp] fix eslint error @typescript-eslint/no-require-imports
