+++ sdk/core/core-amqp/test/context.spec.ts
@@ -5,7 +5,6 @@ import * as chai from "chai";
 const should = chai.should();
 import { CbsClient, ConnectionConfig, ConnectionContextBase } from "../src";
 import { Connection } from "rhea-promise";
-import { isNode } from "../src/util/utils";
 
 describe("ConnectionContextBase", function() {
   it("should be created with required parameters", function(done) {
@@ -249,50 +248,6 @@ describe("ConnectionContextBase", function() {
     );
     context.cbsSession.should.instanceOf(CbsClient);
   });
-
-  if (isNode) {
-    it("should accept a websocket constructor in Node", async () => {
-      const connectionString =
-        "Endpoint=sb://hostname.servicebus.windows.net/;SharedAccessKeyName=sakName;SharedAccessKey=sak;EntityPath=ep";
-      const path = "mypath";
-      const config = ConnectionConfig.create(connectionString, path);
-
-      config.webSocket = require("ws");
-      config.webSocketEndpointPath = "/ws";
-
-      const context = ConnectionContextBase.create({
-        config: config,
-        connectionProperties: {
-          product: "MSJSClient",
-          userAgent: "/js-amqp-client",
-          version: "1.0.0"
-        }
-      });
-
-      should.exist(context);
-    });
-  } else {
-    it("should default to using a websocket in the browser", async () => {
-      const connectionString =
-        "Endpoint=sb://hostname.servicebus.windows.net/;SharedAccessKeyName=sakName;SharedAccessKey=sak;EntityPath=ep";
-      const path = "mypath";
-      const config = ConnectionConfig.create(connectionString, path);
-
-      config.webSocketEndpointPath = "/ws";
-
-      const context = ConnectionContextBase.create({
-        config: config,
-        connectionProperties: {
-          product: "MSJSClient",
-          userAgent: "/js-amqp-client",
-          version: "1.0.0"
-        }
-      });
-
-      should.exist(context);
-      should.exist(context.connection.options.connection_details);
-    });
-  }
 
   it("Throws error if user-agent string length is greater than 512 characters", function(done) {
     const connectionString =
